refactor(frontend): move MobileMenu to Layoute ml-155
