+++ apps/frontend/src/libs/components/layout/styles.module.css
@@ -36,4 +36,9 @@
 	.layout__main {
 		padding: 20px;
 	}
+
+	.layout--no-scroll {
+		overflow: hidden;
+		touch-action: none;
+	}
 }
